requirements.txt
hydra-core==1.1.0
open_clip_torch==2.10.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
matplotlib==3.6.0
numpy
Pillow==9.2.0
scikit-learn==1.1.3
scipy==1.9.3
tqdm==4.64.1
pytorch-lightning==1.9.0
word2number==1.1
scipy